# docs/contracts/openapi_v1.yaml
openapi: 3.1.0
info:
  title: Finwealth API
  version: 0.1.0
  description: >
    Draft HTTP API for the personal asset / investment / liability control
    panel. This API never executes transfers, places orders, or lets AI write
    directly to the confirmed ledger.
servers:
  - url: /v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login with username, password, and device name.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Tokens for this device.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Refresh access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Refreshed tokens.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
...
  /auth/devices/{deviceId}/revoke:
    post:
      tags: [Auth]
      summary: Revoke a device.
      parameters:
        - $ref: "#/components/parameters/deviceId"
      responses:
        "204":
          description: Device revoked.
...
  /accounts:
    get:
      tags: [Accounts]
      summary: List accounts.
      responses:
        "200":
          description: Accounts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListResponse"
    post:
      tags: [Accounts]
      summary: Create account.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountInput"
      responses:
        "200":
          description: Created account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
...
  /accounts/{accountId}:
    get:
      tags: [Accounts]
      summary: Get account.
      parameters:
        - $ref: "#/components/parameters/accountId"
      responses:
        "200":
          description: Account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
    patch:
      tags: [Accounts]
      summary: Update account.
      parameters:
        - $ref: "#/components/parameters/accountId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Updated account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
  /accounts/{accountId}/archive:
    post:
      tags: [Accounts]
      summary: Archive account without deleting ledger history.
      parameters:
        - $ref: "#/components/parameters/accountId"
        - $ref: "#/components/parameters/idempotencyKey"
      responses:
        "200":
          description: Archived account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
  /accounts/{accountId}/holdings:
    get:
      tags: [Portfolio]
      summary: List holdings by account.
      parameters:
        - $ref: "#/components/parameters/accountId"
      responses:
        "200":
          description: Holdings in the account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HoldingListResponse"
...
  /movements:
    get:
      tags: [Movements]
      summary: List movements.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/MovementStatus"
      responses:
        "200":
          description: Movements.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementListResponse"
  /movements/drafts:
    post:
      tags: [Movements]
      summary: Create a manual draft; does not affect confirmed net worth.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMovementDraftInput"
      responses:
        "200":
          description: Draft movement.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementResponse"
  /movements/{movementId}:
    get:
      tags: [Movements]
      summary: Get movement.
      parameters:
        - $ref: "#/components/parameters/movementId"
      responses:
        "200":
          description: Movement.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementResponse"
  /movements/{movementId}/submit-review:
    post:
      tags: [Movements]
      summary: Submit a draft movement for review.
      parameters:
        - $ref: "#/components/parameters/movementId"
        - $ref: "#/components/parameters/idempotencyKey"
      responses:
        "200":
          description: Atomic group proposal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAtomicGroupResponse"
  /movements/corrections:
    post:
      tags: [Movements]
      summary: Create a correction proposal for confirmed ledger data.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCorrectionInput"
      responses:
        "200":
          description: Correction atomic group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAtomicGroupResponse"
  /atomic-groups/{atomicGroupId}/confirm:
    post:
      tags: [Movements]
      summary: Confirm an atomic group as one transaction boundary.
      parameters:
        - $ref: "#/components/parameters/atomicGroupId"
        - $ref: "#/components/parameters/idempotencyKey"
      responses:
        "200":
          description: Confirm result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmResultResponse"
  /atomic-groups/{atomicGroupId}/reject:
    post:
      tags: [Movements]
      summary: Reject an atomic group.
      parameters:
        - $ref: "#/components/parameters/atomicGroupId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequest"
      responses:
        "204":
          description: Rejected.
  /dca/plans:
    get:
      tags: [DCA]
      summary: List DCA plans.
      responses:
        "200":
          description: DCA plans.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcaPlanListResponse"
    post:
      tags: [DCA]
      summary: Create DCA plan.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDcaPlanInput"
      responses:
        "200":
          description: DCA plan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcaPlanResponse"
  /dca/plans/{planId}:
    patch:
      tags: [DCA]
      summary: Update DCA plan.
      parameters:
        - $ref: "#/components/parameters/planId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: DCA plan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcaPlanResponse"
...
  /dca/reminders/{reminderId}/mark-executed-as-proposal:
    post:
      tags: [DCA]
      summary: Create a proposal that records the user-executed DCA; never places an order.
      parameters:
        - $ref: "#/components/parameters/reminderId"
        - $ref: "#/components/parameters/idempotencyKey"
      responses:
        "200":
          description: Candidate atomic group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAtomicGroupResponse"
  /dca/reminders/{reminderId}/skip:
    post:
      tags: [DCA]
      summary: Skip this DCA reminder.
      parameters:
        - $ref: "#/components/parameters/reminderId"
        - $ref: "#/components/parameters/idempotencyKey"
      responses:
        "200":
          description: Updated reminder.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcaReminderResponse"
  /dca/reminders/{reminderId}/snooze:
    post:
      tags: [DCA]
      summary: Snooze this DCA reminder.
      parameters:
        - $ref: "#/components/parameters/reminderId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [until]
              properties:
                until:
                  $ref: "#/components/schemas/ISODateTime"
      responses:
        "200":
          description: Updated reminder.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcaReminderResponse"
  /ai/proposals/from-text:
    post:
      tags: [AI Proposals]
      summary: Create AI proposal from text.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiTextInput"
      responses:
        "200":
          description: AI proposal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProposalResponse"
  /ai/proposals/from-image:
    post:
      tags: [AI Proposals]
      summary: Create AI proposal from image evidence.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiImageInput"
      responses:
        "200":
          description: AI proposal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProposalResponse"
  /ai/proposals/from-csv:
    post:
      tags: [AI Proposals]
      summary: Create AI proposal from CSV import.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiCsvInput"
      responses:
        "200":
          description: AI proposal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProposalResponse"
...
  /ai/proposals/{proposalId}:
    get:
      tags: [AI Proposals]
      summary: Get AI proposal.
      parameters:
        - $ref: "#/components/parameters/proposalId"
      responses:
        "200":
          description: AI proposal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProposalResponse"
  /ai/atomic-groups/{atomicGroupId}/approve:
    post:
      tags: [AI Proposals]
      summary: Approve atomic group after validation.
      parameters:
        - $ref: "#/components/parameters/atomicGroupId"
        - $ref: "#/components/parameters/idempotencyKey"
      responses:
        "200":
          description: Confirm result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmResultResponse"
  /ai/atomic-groups/{atomicGroupId}/reject:
    post:
      tags: [AI Proposals]
      summary: Reject AI atomic group.
      parameters:
        - $ref: "#/components/parameters/atomicGroupId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequest"
      responses:
        "204":
          description: Rejected.
  /ai/atomic-groups/{atomicGroupId}/edit:
    post:
      tags: [AI Proposals]
      summary: Edit atomic group and re-run validation.
      parameters:
        - $ref: "#/components/parameters/atomicGroupId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Edited atomic group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAtomicGroupResponse"
...
  /quotes/refresh:
    post:
      tags: [Quotes]
      summary: Refresh quotes and FX rates.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRefreshRequest"
      responses:
        "200":
          description: Refresh result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRefreshResultResponse"
  /instruments/{instrumentId}/historical-prices:
    get:
      tags: [Quotes]
      summary: Get historical prices, capped to one year for MVP.
      parameters:
        - $ref: "#/components/parameters/instrumentId"
        - name: from
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ISODate"
        - name: to
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ISODate"
      responses:
        "200":
          description: Historical prices.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoricalPriceListResponse"
...
  /snapshots:
    get:
      tags: [Snapshots]
      summary: List snapshots.
      parameters:
        - name: from
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ISODate"
        - name: to
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ISODate"
      responses:
        "200":
          description: Snapshots.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetWorthSnapshotListResponse"
  /snapshots/manual:
    post:
      tags: [Snapshots]
      summary: Create a manual snapshot.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  enum: [baseline, manual_refresh]
      responses:
        "200":
          description: Snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetWorthSnapshotResponse"
  /snapshots/invalidate:
    post:
      tags: [Snapshots]
      summary: Invalidate snapshots after ledger changes.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "204":
          description: Invalidated.
  /categories:
    get:
      tags: [Taxonomy]
      summary: List categories.
      responses:
        "200":
          description: Categories.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryListResponse"
    post:
      tags: [Taxonomy]
      summary: Create category.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
      responses:
        "200":
          description: Category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
  /categories/{categoryId}:
    patch:
      tags: [Taxonomy]
      summary: Update category.
      parameters:
        - $ref: "#/components/parameters/categoryId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
  /counterparties:
    get:
      tags: [Taxonomy]
      summary: List counterparties.
      responses:
        "200":
          description: Counterparties.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CounterpartyListResponse"
    post:
      tags: [Taxonomy]
      summary: Create counterparty.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Counterparty"
      responses:
        "200":
          description: Counterparty.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CounterpartyResponse"
  /counterparties/{counterpartyId}:
    patch:
      tags: [Taxonomy]
      summary: Update counterparty.
      parameters:
        - $ref: "#/components/parameters/counterpartyId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Counterparty.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CounterpartyResponse"
  /counterparties/merge-proposal:
    post:
      tags: [Taxonomy]
      summary: Create counterparty merge proposal; never auto-merges.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceCounterpartyIds, targetDisplayName]
              properties:
                sourceCounterpartyIds:
                  type: array
                  items:
                    $ref: "#/components/schemas/ID"
                targetDisplayName:
                  type: string
      responses:
        "200":
          description: Merge proposal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAtomicGroupResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
    accountId:
      name: accountId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ID"
...
    deviceId:
      name: deviceId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ID"
  schemas:
...
    ISODate:
      type: string
      format: date
...
    LoginRequest:
      type: object
      required: [username, password, deviceName]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        deviceName:
          type: string
    AuthTokenResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          const: true
        data:
          type: object
          required: [accessToken, refreshToken, expiresAt, deviceId]
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            expiresAt:
              $ref: "#/components/schemas/ISODateTime"
            deviceId:
              $ref: "#/components/schemas/ID"
...
    AccountListResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          const: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Account"
...